fix: bin
